% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynSysSimR-package.R
\docType{package}
\name{DynSysSimR-package}
\alias{DynSysSimR}
\alias{DynSysSimR-package}
\title{DynSysSimR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
